ClassCircularityError: Ensure transformer does not depend kotlin, java or CHR classes
Compose Hot Reload uses [java.lang.instrument.ClassFileTransformer]s to hook into the JVMs class loading.
When a class is loaded, we might enqueue the code for bytecode analysis or even transform the code.
However, the transformers themselves do require some code to be loaded.
If the transformer requires e.g. 'kotlin.List' to work, but loading 'kotlin.List' requires running the transformer,
then we end up with our infamous 'ClassCircularityError'.
To avoid this, the transformer is *not allowed* to transform any code it itself depends upon.
Note: There is also a [isIgnored] property which even allows external systems to extend the list of 'ignored' classes.