Orchestration: Load encoders from the context of the key
This shall allow more complex systems to extend the orchestration where
states (and their corresponding keys) might not be loaded at the same level
where the orchestration is loaded.

Since the orchestration is required by the agent, it is typically loaded
at system level, whereas the UI or other parts might be loaded
in child class loaders.

Providing a state from such a classloader also requires loading
the encoder from a suitable context.